moved fetches to seperate file
